# service/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenID Federation Registry API",
    "description": "REST API for managing OpenID Federation entities, policies, modules, and trust marks. This API provides endpoints for creating, reading, updating, and deleting federation entities, policies, trust anchors, resolvers, trust marks, and trust mark subjects.",
    "contact": {
      "name": "Sweden Connect",
      "url": "https://www.swedenconnect.se",
      "email": "support@swedenconnect.se"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.5.10-SNAPSHOT"
  },
  "servers": [
    {
      "url": "https://registry.local.swedenconnect.se",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/registration-admin/v1/{tenant}/{orgNumber}/{registrationId}/steps/{stepIndex}/approve": {
      "post": {
        "tags": [
          "RegistrationAdmin"
        ],
        "summary": "Approve a specific pending pipeline step and resume execution",
        "operationId": "approveStep",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "registrationId",
            "in": "path",
            "description": "Registration ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "stepIndex",
            "in": "path",
            "description": "Step index",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                }
              }
            }
          }
        }
      }
    },
    "/registration-admin/v1/{tenant}/{orgNumber}/{registrationId}/reject": {
      "post": {
        "tags": [
          "RegistrationAdmin"
        ],
        "summary": "Reject a pending registration request",
        "operationId": "reject",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "registrationId",
            "in": "path",
            "description": "Registration ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                }
              }
            }
          }
        }
      }
    },
...
    "/registration/v1/{tenant}/{orgNumber}": {
      "get": {
        "tags": [
          "FederationRegistration"
        ],
        "summary": "List all registration records for current organization",
        "operationId": "listRegistrations",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Registration"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/registration-admin/v1/{tenant}/{orgNumber}": {
      "get": {
        "tags": [
          "RegistrationAdmin"
        ],
        "summary": "List all registration records for current organization",
        "operationId": "listRegistrations_1",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Registration"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/registration-admin/v1/{tenant}/{orgNumber}/{registrationId}": {
      "get": {
        "tags": [
          "RegistrationAdmin"
        ],
        "summary": "Get a single registration by ID",
        "operationId": "getById_1",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "registrationId",
            "in": "path",
            "description": "Registration ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                }
              }
            }
          }
        }
      }
    },
    "/registration-admin/v1/{tenant}/{orgNumber}/count": {
      "get": {
        "tags": [
          "RegistrationAdmin"
        ],
        "summary": "Count unhandled PENDING registrations for an intermediate",
        "operationId": "countPending",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "taimId",
            "in": "query",
            "description": "Intermediate ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextDiffEntry": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Context key that changed"
          },
          "changeType": {
            "type": "string",
            "description": "Type of change: ADDED, CHANGED, or REMOVED"
          },
          "before": {
            "type": "string",
            "description": "Value before execution, null for ADDED entries"
          },
          "after": {
            "type": "string",
            "description": "Value after execution, null for REMOVED entries"
          }
        }
      },
      "Registration": {
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid",
            "description": "Id of this specific registration",
            "readOnly": true
          },
          "joinId": {
            "type": "string",
            "format": "uuid",
            "description": "Join ID",
            "readOnly": true
          },
          "entityIdentifier": {
            "type": "string",
            "description": "EntityId for the requested registration",
            "example": "https://example.com/service",
            "readOnly": true
          },
          "intermediateEntityId": {
            "type": "string",
            "description": "EntityId of the Intermediate that the registration is made on",
            "example": "https://example.com/entity",
            "readOnly": true
          },
          "isHosted": {
            "type": "boolean",
            "description": "True if the entity's metadata is hosted in this registry",
            "readOnly": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "description": "Hosted metadata, present when isHosted is true",
            "readOnly": true
          },
          "registrationType": {
            "type": "string",
            "description": "Registration type: SUBORDINATE or TRUST_MARK_SUBORDINATE",
            "enum": [
              "SUBORDINATE",
              "TRUST_MARK_SUBORDINATE"
            ],
            "readOnly": true
          },
          "statusFedreg": {
            "type": "string",
            "description": "Current federation registration status",
            "enum": [
              "STARTED",
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED"
            ],
            "readOnly": true
          },
          "rejectionReason": {
            "type": "string",
            "description": "Reason for rejection, present when status_fedreg is DENY",
            "readOnly": true
          },
          "statusTrustmarks": {
            "type": "array",
            "description": "Trustmark status per requested trustmark",
            "items": {
              "$ref": "#/components/schemas/TrustmarkRegistration"
            },
            "readOnly": true
          },
          "tags": {
            "type": "array",
            "description": "Tags describing the entity type, e.g. OIDC, SAML, RP, OP, IDP, SP",
            "example": [
              "OIDC",
              "RP"
            ],
            "items": {
              "type": "string",
              "enum": [
                "OIDC",
                "SAML",
                "HOSTED",
                "FED",
                "RP",
                "OP",
                "AS",
                "OAC",
                "ORS",
                "SAML_IDP",
                "SAML_SP",
                "SAML_RP"
              ]
            },
            "readOnly": true
          },
          "organizationName": {
            "type": "string",
            "description": "Name of the organization that submitted the registration",
            "readOnly": true
          },
          "subordinateEntityId": {
            "type": "string",
            "description": "Entity ID of the subordinate subject, populated for TRUST_MARK_SUBORDINATE registrations",
            "readOnly": true
          },
          "successful": {
            "type": "boolean",
            "description": "True if the pipeline completed without any step failures"
          },
          "steps": {
            "type": "array",
            "description": "Ordered list of step execution records",
            "items": {
              "$ref": "#/components/schemas/StepExecutionRecord"
            }
          }
        }
      },
      "StepExecutionRecord": {
        "type": "object",
        "properties": {
          "stepName": {
            "type": "string",
            "description": "Display name of the step",
            "example": "LoadEntityConfigurationStep"
          },
          "status": {
            "type": "string",
            "description": "Step execution outcome",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "description": "Human-readable summary of the step result"
          },
          "issues": {
            "type": "array",
            "description": "Validation issues found during this step",
            "items": {
              "$ref": "#/components/schemas/StepIssue"
            }
          },
          "contextDiff": {
            "type": "array",
            "description": "Changes made to the pipeline context by this step",
            "items": {
              "$ref": "#/components/schemas/ContextDiffEntry"
            }
          }
        }
      },
      "StepIssue": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Field that caused the issue",
            "example": "InternalPreRegistrationStep.status"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the issue"
          },
          "severity": {
            "type": "string",
            "description": "Severity of the issue",
            "example": "ERROR"
          }
        }
      },
      "TrustmarkRegistration": {
        "type": "object",
        "properties": {
          "trustmarkIssuer": {
            "type": "string",
            "description": "Entity identifier of the trustmark issuer",
            "example": "https://fed.swedenconnect.se/tmi"
          },
          "trustmarkStatus": {
            "type": "array",
            "description": "Trustmarks with status",
            "items": {
              "$ref": "#/components/schemas/TrustmarkStatus"
            }
          }
        }
      },
      "TrustmarkStatus": {
        "type": "object",
        "properties": {
          "trustmarkType": {
            "type": "string",
            "description": "Trustmark type identifier (URI)",
            "example": "https://trust.example.com/tm/loa3"
          },
          "status": {
            "type": "string",
            "description": "Current status of this trustmark",
            "enum": [
              "STARTED",
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      },
...
      "RejectRegistrationRequest": {
        "type": "object",
        "properties": {
          "rejectionReason": {
            "type": "string",
            "description": "Human-readable explanation shown to the applicant",
            "example": "Entity configuration could not be fetched from the provided entity ID."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token obtained from authentication endpoint",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}